consumer/requirements.txt
kafka-python
psycopg2-binary
python-dotenv
boto3
requests
